requirements.txt
backports-abc==0.5
backports.shutil-get-terminal-size==1.0.0
certifi==2017.7.27.1
decorator==4.1.2
Django==1.11.6
django-cors-headers==2.1.0
django-filter==1.0.4
django-jsoneditor==0.0.12
djangorestframework==3.7.0
docutils==0.14
enum34==1.1.6
ipython==5.5.0
ipython-genutils==0.2.0
jsonfield==2.0.2
lockfile==0.12.2
luigi==2.7.1
Markdown==2.6.9
MySQL-python==1.2.5
numpy==1.13.3
packaging==16.8
pandas==0.20.3
pathlib2==2.3.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
psutil==5.4.0
ptyprocess==0.5.2
Pygments==2.2.0
pyparsing==2.2.0
python-daemon==2.1.2
python-dateutil==2.6.1
pytz==2017.2
scandir==1.6
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
SQLAlchemy==1.1.14
tornado==4.5.2
traitlets==4.3.2
wcwidth==0.1.7
